chore(sdk): Move all `RoomEventCache` types from `mod.rs` to `room/mod.rs`.
